only move if exists

+++ setup.sh
@@ -24,7 +24,9 @@ do
         echo "${DOTFILES_DIR}/${file} does not exist"
 
     elif ! cmp --silent "~/${file}" "${DOTFILES_DIR}/${file}"; then
-        mv "~/${file}" "${BACKUP_DIR}/${file}"
+        if ! [[ -f "${file}" || -d "${file}" ]]; then
+            mv "~/${file}" "${BACKUP_DIR}/${file}"
+        fi
         ln -s "${DOTFILES_DIR}/${file}" "${file}"
 
     fi
